Refactor DfpAgent lifecycle
